refactor: Move implementation to poly.rs.
